+++ be/issue-tracker/src/main/java/com/issuetracker/file/Repository/FileRepository.java
@@ -1,6 +1,6 @@
 package com.issuetracker.file.Repository;
 
-import com.issuetracker.file.domain.File;
+import com.issuetracker.file.entity.File;
 import org.springframework.data.repository.CrudRepository;
 
 public interface FileRepository extends CrudRepository<File, Long> {
